REFPLTV-1146 Bring up 64bit support in Rpi4 (for both kernel and user space).
Reason for change: to match datatype compatibilty
Test Procedure: Build should be successful. display settings should build fine.
Risks: Low/None.

+++ dsDisplay.c
@@ -134,7 +134,7 @@ dsError_t dsGetDisplay(dsVideoPortType_t m_vType, int index, int *handle)
     }
 
     if (ret == dsERR_NONE) {
-        *handle = (int)&_handles[m_vType][index];
+        *handle = (intptr_t)&_handles[m_vType][index];
 
     } else {
     }
